Delete yarn. lock, Create a separate component for to-do list and Add to-do list form
